Unlimmited blacksmith, some more food from rats and crabs and no more negative amounts of gold

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/actors/mobs/Mob.java
@@ -867,7 +867,7 @@ public abstract class Mob extends Char {
 	public void rollToDropLoot(){
 		//Al enemies now always drop gold
 		Gold gold = new Gold(Random.NormalIntRange(1, maxLvl));
-		Dungeon.level.drop(gold, pos).sprite.drop();
+		if (gold.quantity() > 0)Dungeon.level.drop(gold, pos).sprite.drop();
 
 		MasterThievesArmband.StolenTracker stolen = buff(MasterThievesArmband.StolenTracker.class);
 		if (stolen == null || !stolen.itemWasStolen()) {
